Started Business Logic

+++ src/main/java/com/autogenfoodplaceapp/autogenfoodplaceapp/Repository/FoodPlaceRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FoodPlaceRepository extends JpaRepository<FoodPlace, Long> {
+public interface FoodPlaceRepository extends JpaRepository<FoodPlace, Integer> {
 }
